requirements.txt
requests~=2.23.0
discord.py
dblpy